+++ source/enumerific/extensible.py
@@ -9,6 +9,7 @@ from enumerific.exceptions import (
     EnumerationError,
     EnumerationOptionError,
     EnumerationSubclassingError,
+    EnumerationExtensibilityError,
     EnumerationNonUniqueError,
 )
 
@@ -196,10 +197,12 @@ class EnumerationConfiguration(object):
     """The EnumerationConfiguration class holds the Enumeration configuration options"""
 
     _unique: bool = None
-    _aliased: bool = False
+    _aliased: bool = None
+    _backfill: bool = None
     _overwritable: bool = None
     _removable: bool = None
     _subclassable: bool = None
+    _extensible: bool = None
     _raises: bool = None
     _flags: bool = None
     _start: int = None
@@ -211,9 +214,11 @@ class EnumerationConfiguration(object):
         self,
         unique: bool = None,
         aliased: bool = None,
+        backfill: bool = None,
         overwritable: bool = None,
         removable: bool = None,
         subclassable: bool = None,
+        extensible: bool = None,
         raises: bool = None,
         flags: bool = None,
         start: int = None,
@@ -223,9 +228,11 @@ class EnumerationConfiguration(object):
     ):
         self.unique = unique
         self.aliased = aliased
+        self.backfill = backfill
         self.overwritable = overwritable
         self.removable = removable
         self.subclassable = subclassable
+        self.extensible = extensible
         self.raises = raises
         self.flags = flags
         self.start = start
@@ -237,9 +244,11 @@ class EnumerationConfiguration(object):
         return [
             "unique",
             "aliased",
+            "backfill",
             "overwritable",
             "removable",
             "subclassable",
+            "extensible",
             "raises",
             "flags",
             "start",
@@ -311,6 +320,20 @@ class EnumerationConfiguration(object):
         self._aliased = aliased
 
     @property
+    def backfill(self) -> bool | None:
+        return self._backfill
+
+    @backfill.setter
+    def backfill(self, backfill: bool | None):
+        if backfill is None:
+            pass
+        elif not isinstance(backfill, bool):
+            raise TypeError(
+                "The 'backfill' argument, if specified, must have a boolean value!"
+            )
+        self._backfill = backfill
+
+    @property
     def overwritable(self) -> bool | None:
         return self._overwritable
 
@@ -351,6 +374,20 @@ class EnumerationConfiguration(object):
                 "The 'subclassable' argument, if specified, must have a boolean value!"
             )
         self._subclassable = subclassable
+
+    @property
+    def extensible(self) -> bool | None:
+        return self._extensible
+
+    @extensible.setter
+    def extensible(self, extensible: bool | None):
+        if extensible is None:
+            pass
+        elif not isinstance(extensible, bool):
+            raise TypeError(
+                "The 'extensible' argument, if specified, must have a boolean value!"
+            )
+        self._extensible = extensible
 
     @property
     def raises(self) -> bool | None:
@@ -463,8 +500,10 @@ class EnumerationMetaClass(type):
         bases: tuple[type],
         unique: bool = None,
         aliased: bool = None,
+        backfill: bool = None,
         overwritable: bool = None,
         subclassable: bool = None,
+        extensible: bool = None,
         removable: bool = None,
         raises: bool = None,
         flags: bool = None,
@@ -481,14 +520,16 @@ class EnumerationMetaClass(type):
         any other keyword arguments that are included in the class signature call."""
 
         logger.debug(
-            "[EnumerationMetaClass] %s.__prepare__(name: %s, bases: %s, unique: %s, aliased: %s, overwritable: %s, subclassable: %s, removable: %s, raises: %s, flags: %s, start: %s, steps: %s, times: %s, typecast: %s, kwargs: %s)",
+            "[EnumerationMetaClass] %s.__prepare__(name: %s, bases: %s, unique: %s, aliased: %s, backfill: %s, overwritable: %s, subclassable: %s, extensible: %s, removable: %s, raises: %s, flags: %s, start: %s, steps: %s, times: %s, typecast: %s, kwargs: %s)",
             name,
             name,
             bases,
             unique,
             aliased,
+            backfill,
             overwritable,
             subclassable,
+            extensible,
             removable,
             raises,
             flags,
@@ -581,8 +622,10 @@ class EnumerationMetaClass(type):
         *args,
         unique: bool = None,  # True
         aliased: bool = None,  # False
+        backfill: bool = None,  # False
         overwritable: bool = None,  # False
         subclassable: bool = None,  # True
+        extensible: bool = None,  # True
         removable: bool = None,  # False
         raises: bool = None,  # False
         flags: bool = None,  # False
@@ -613,6 +656,13 @@ class EnumerationMetaClass(type):
                 "The 'aliased' argument, if specified, must have a boolean value!"
             )
 
+        if backfill is None:
+            pass
+        elif not isinstance(backfill, bool):
+            raise TypeError(
+                "The 'backfill' argument, if specified, must have a boolean value!"
+            )
+
         if overwritable is None:
             pass
         elif not isinstance(overwritable, bool):
@@ -625,6 +675,13 @@ class EnumerationMetaClass(type):
         elif not isinstance(subclassable, bool):
             raise TypeError(
                 "The 'subclassable' argument, if specified, must have a boolean value!"
+            )
+
+        if extensible is None:
+            pass
+        elif not isinstance(extensible, bool):
+            raise TypeError(
+                "The 'extensible' argument, if specified, must have a boolean value!"
             )
 
         if removable is None:
@@ -679,8 +736,10 @@ class EnumerationMetaClass(type):
         configuration = EnumerationConfiguration(
             unique=unique,
             aliased=aliased,
+            backfill=backfill,
             overwritable=overwritable,
             subclassable=subclassable,
+            extensible=extensible,
             removable=removable,
             raises=raises,
             flags=flags,
@@ -747,10 +806,16 @@ class EnumerationMetaClass(type):
                             " >>> aliased      => %s", base_configuration.aliased
                         )
                         logger.debug(
+                            " >>> backfill     => %s", base_configuration.backfill
+                        )
+                        logger.debug(
                             " >>> overwritable => %s", base_configuration.overwritable
                         )
                         logger.debug(
                             " >>> subclassable => %s", base_configuration.subclassable
+                        )
+                        logger.debug(
+                            " >>> extensible   => %s", base_configuration.extensible
                         )
                         logger.debug(
                             " >>> removable    => %s", base_configuration.removable
@@ -774,9 +839,12 @@ class EnumerationMetaClass(type):
                             " >>> typecast     => %s", base_configuration.typecast
                         )
 
-                        if base_configuration.subclassable is False:
+                        if (
+                            base_configuration.subclassable is False
+                            or base_configuration.extensible is False
+                        ):
                             raise EnumerationSubclassingError(
-                                "The '%s' enumeration class cannot be subclassed when the keyword argument 'subclassable=False' was passed to the class constructor!"
+                                "The '%s' enumeration class cannot be subclassed when the keyword arguments 'subclassable=False' or 'extensible=False` are passed to the class constructor!"
                                 % (base.__name__)
                             )
 
@@ -792,12 +860,19 @@ class EnumerationMetaClass(type):
                             " >>> (updated) aliased      => %s", configuration.aliased
                         )
                         logger.debug(
+                            " >>> (updated) backfill     => %s", configuration.backfill
+                        )
+                        logger.debug(
                             " >>> (updated) overwritable => %s",
                             configuration.overwritable,
                         )
                         logger.debug(
                             " >>> (updated) subclassable => %s",
                             configuration.subclassable,
+                        )
+                        logger.debug(
+                            " >>> (updated) extensible => %s",
+                            configuration.extensible,
                         )
                         logger.debug(
                             " >>> (updated) removable    => %s", configuration.removable
@@ -824,7 +899,6 @@ class EnumerationMetaClass(type):
                     # logger.debug(" >>> found base (%s) that is an instance of EnumerationMetaClass and a subclass of Enumeration" % (base))
 
                     if not (base is Enumeration or Enumeration in base.__bases__):
-                        # enumerations = base._enumerations  # reference to the _enumerations dictionary
                         _enumerations = base._enumerations
 
                     logger.debug("  >>> enumerations => %s" % (base._enumerations))
@@ -846,8 +920,10 @@ class EnumerationMetaClass(type):
         configuration.defaults(
             unique=True,
             aliased=False,
+            backfill=False,
             overwritable=False,
             subclassable=True,
+            extensible=True,
             removable=False,
             raises=False,
             flags=False,
@@ -859,12 +935,14 @@ class EnumerationMetaClass(type):
 
         logger.debug(" >>> (after defaults) unique       => %s", configuration.unique)
         logger.debug(" >>> (after defaults) aliased      => %s", configuration.aliased)
+        logger.debug(" >>> (after defaults) backfill     => %s", configuration.backfill)
         logger.debug(
             " >>> (after defaults) overwritable => %s", configuration.overwritable
         )
         logger.debug(
             " >>> (after defaults) subclassable => %s", configuration.subclassable
         )
+        logger.debug(" >>> (after defaults) extensible => %s", configuration.extensible)
         logger.debug(
             " >>> (after defaults) removable    => %s", configuration.removable
         )
@@ -998,28 +1076,20 @@ class EnumerationMetaClass(type):
 
         logger.debug(" >>> bases         => %s", [base for base in bases])
 
-        # if "EnumerationInteger" in globals():
-        #     if EnumerationInteger in bases and EnumerationFlag in bases:
-        #         bases = tuple([base for base in bases if not EnumerationInteger])
-
         args: tuple[object] = (name, bases, attributes)
 
         # Create the new enumeration class instance
         instance = super().__new__(cls, *args, **kwargs)
-
-        # logger.debug(
-        # " >>> metaclass => %s (base: %s, type: %s, bases: %s)\n"
-        # % (enumclass, instance, type(instance), instance.__bases__)
-        # )
-        # logger.debug(" >>> metaclass => %s (base: %s, type: %s, bases: %s)\n" % (enumclass, instance, type(instance), instance.__bases__))
 
         logger.debug(" >>> baseclass     => %s", baseclass)
         logger.debug(" >>> instance      => %s", instance)
 
         logger.debug(" >>> unique        => %s", configuration.unique)
         logger.debug(" >>> aliased       => %s", configuration.aliased)
+        logger.debug(" >>> backfill      => %s", configuration.backfill)
         logger.debug(" >>> overwritable  => %s", configuration.overwritable)
         logger.debug(" >>> subclassable  => %s", configuration.subclassable)
+        logger.debug(" >>> extensible    => %s", configuration.extensible)
         logger.debug(" >>> removable     => %s", configuration.removable)
         logger.debug(" >>> raises        => %s", configuration.raises)
         logger.debug(" >>> flags         => %s", configuration.flags)
@@ -1048,7 +1118,16 @@ class EnumerationMetaClass(type):
         )
 
         if isinstance(base_enumerations := attributes.get("base_enumerations"), dict):
-            self._enumerations: dict[str, Enumeration] = base_enumerations
+            if (
+                self._configuration.backfill is True
+                and self._configuration.extensible is True
+            ):
+                self._enumerations: dict[str, Enumeration] = base_enumerations
+            else:
+                self._enumerations: dict[str, Enumeration] = {}
+
+                for enumeration_name, enumeration in base_enumerations.items():
+                    self._enumerations[enumeration_name] = enumeration
         else:
             self._enumerations: dict[str, Enumeration] = {}
 
@@ -1131,16 +1210,19 @@ class EnumerationMetaClass(type):
             )
 
     def __dir__(self) -> list[str]:
-        members: list[str] = []
-
-        for name, enumeration in self._enumerations.items():
-            members.append(name)
+        members: set[str] = set()
 
         for member in object.__dir__(self):
             if member.startswith("_") or member in self._special:
-                members.append(member)
+                members.add(member)
 
-        return members
+        for name, enumeration in self._enumerations.items():
+            members.add(name)
+
+        for member in vars(self):
+            members.add(member)
+
+        return list(members)
 
     def __contains__(self, other: Enumeration | object) -> bool:
         logger.debug(
@@ -1326,6 +1408,12 @@ class EnumerationMetaClass(type):
             value,
         )
 
+        if self.configuration.extensible is False:
+            raise EnumerationExtensibilityError(
+                "The '%s' enumeration class has been configured to prevent extensibility, so cannot be extended with new options either through registration or subclassing, so the '%s' option cannot be registered!"
+                % (self.__name__, name)
+            )
+
         if self.configuration.overwritable is False and name in self._enumerations:
             raise EnumerationNonUniqueError(
                 "The '%s' enumeration class already has an option named '%s', so a new option with the same name cannot be created unless the 'overwritable=True' argument is passed during class construction!"
@@ -1430,6 +1518,14 @@ class EnumerationMetaClass(type):
                     )
                 )
 
+        # When an enumeration option is reconciled, it may be defined in another class
+        # but have been accessed through a subclass; in order for attribute lookups to
+        # work within the subclass, we need to provide the current lookup context to the
+        # reconciled enumeration option, so that any attribute access on this object can
+        # perform their lookup in the correct part of the class hierarchy
+        if isinstance(reconciled, Enumeration):
+            reconciled._context = self
+
         return reconciled
 
     def validate(self, value: Enumeration | object = None, name: str = None) -> bool:
@@ -1449,6 +1545,7 @@ class Enumeration(metaclass=EnumerationMetaClass):
     """The Enumeration class is the subclass of all enumerations and their subtypes."""
 
     _metaclass: EnumerationMetaClass = None
+    _context: EnumerationMetaClass = None
     _enumeration: Enumeration = None
     _enumerations: dict[str, Enumeration] = None
     _annotations: anno = None
@@ -1609,6 +1706,8 @@ class Enumeration(metaclass=EnumerationMetaClass):
             return self._enumerations[name]
         elif self._annotations and name in self._annotations:
             return self._annotations[name]
+        elif self._context and name in dir(self._context):
+            return object.__getattribute__(self._context, name)
         else:
             # EnumerationOptionError subclasses AttributeError so we adhere to convention
             raise EnumerationOptionError(